+++ src/protocol.rs
@@ -3,5 +3,5 @@ pub mod frame;
 pub mod types;
 
 pub use crc8::calc_crc8;
-pub use frame::{Frame, FrameBuilder};
+pub use frame::{Frame, FrameBuilder, Response};
 pub use types::{Command, ErrorCode};